basic server support for unit tests

+++ Sources/Security/Security.swift
@@ -22,7 +22,7 @@
 
 import Foundation
 
-public enum SecurityErrorCode: Int {
+public enum SecurityErrorCode: UInt16 {
     case acceptFailed = 1
     case pinningFailed = 2
 }
